mediacapture-streams/MediaStreamTrack-getSettings.https.html

<!doctype html>
<html>
<head>
<title>MediaStreamTrack GetSettings</title>
<link rel="author" title="Harald Tveit Alvestrand" href="mailto:hta@google.com"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#widl-MediaStreamTrack-id">
</head>
<body>
<p class="instructions">When prompted, accept to share your video stream.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that settings returned from track.getSettings() can be modified and used to get new user media.</p>

<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  'use strict'

  promise_test(function() {
    const constraints = {video: true, audio: false};

    return navigator.mediaDevices.getUserMedia()
      .then(function(s) {
        const settings1 = s.getVideoTracks()[0].getSettings();
        const videoConstraints = {deviceId: settings1.deviceId};
        return navigator.mediaDevices.getUserMedia({video: videoConstraints, audio: false})
          .then(function(s) {
            const settings2 = s.getVideoTracks()[0].getSettings();
            assert_equals(settings1.deviceId, settings2.deviceId);
          });
      });
  }, 'A device can be opened twice and have the same device ID');

  promise_test(function() {
    const constraints = {video: true, audio: false};

    return navigator.mediaDevices.getUserMedia(constraints)
      .then(function(s) {
        const settings1 = s.getVideoTracks()[0].getSettings();
        const videoConstraints = {deviceId: settings1.deviceId,
                                  width: { exact: settings1.width / 2 }};

        return navigator.mediaDevices.getUserMedia({video: videoConstraints, audio: false})
          .then(function(s) {
            const settings2 = s.getVideoTracks()[0].getSettings();
            assert_equals(settings1.deviceId, settings2.deviceId);
            assert_equals(settings1.width / 2, settings2.width);
          });
      });
  }, 'A device can be opened twice with different resolutions');
</script>
</body>
</html>
